// src/__generated__/graphql.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
};
// ...
export type Film = Node & {
    __typename?: 'Film';
    characterConnection?: Maybe<FilmCharactersConnection>;
    /** The ISO 8601 date format of the time that this resource was created. */
    created?: Maybe<Scalars['String']>;
    /** The name of the director of this film. */
    director?: Maybe<Scalars['String']>;
    /** The ISO 8601 date format of the time that this resource was edited. */
    edited?: Maybe<Scalars['String']>;
    /** The episode number of this film. */
    episodeID?: Maybe<Scalars['Int']>;
    /** The ID of an object */
    id: Scalars['ID'];
    /** The opening paragraphs at the beginning of this film. */
    openingCrawl?: Maybe<Scalars['String']>;
    planetConnection?: Maybe<FilmPlanetsConnection>;
    /** The name(s) of the producer(s) of this film. */
    producers?: Maybe<Array<Maybe<Scalars['String']>>>;
    /** The ISO 8601 date format of film release at original creator country. */
    releaseDate?: Maybe<Scalars['String']>;
    speciesConnection?: Maybe<FilmSpeciesConnection>;
    starshipConnection?: Maybe<FilmStarshipsConnection>;
    /** The title of this film. */
    title?: Maybe<Scalars['String']>;
    vehicleConnection?: Maybe<FilmVehiclesConnection>;
};
// ...
export type Node = {
    /** The id of the object. */
    id: Scalars['ID'];
};

/** A connection to a list of items. */
export type FilmCharactersConnection = {
    __typename?: 'FilmCharactersConnection';
    /**
     * A list of all of the objects returned in the connection. This is a convenience
     * field provided for quickly exploring the API; rather than querying for
     * "{ edges { node } }" when no edge data is needed, this field can be be used
     * instead. Note that when clients like Relay need to fetch the "cursor" field on
     * the edge to enable efficient pagination, this shortcut cannot be used, and the
     * full "{ edges { node } }" version should be used instead.
     */
    characters?: Maybe<Array<Maybe<Person>>>;
    /** A list of edges. */
    edges?: Maybe<Array<Maybe<FilmCharactersEdge>>>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
    /**
     * A count of the total number of objects in this connection, ignoring pagination.
     * This allows a client to fetch the first five objects by passing "5" as the
     * argument to "first", then fetch the total count so it could display "5 of 83",
     * for example.
     */
    totalCount?: Maybe<Scalars['Int']>;
};

/** An individual person or character within the Star Wars universe. */
export type Person = Node & {
    __typename?: 'Person';
    /**
     * The birth year of the person, using the in-universe standard of BBY or ABY -
     * Before the Battle of Yavin or After the Battle of Yavin. The Battle of Yavin is
     * a battle that occurs at the end of Star Wars episode IV: A New Hope.
     */
    birthYear?: Maybe<Scalars['String']>;
    /** The ISO 8601 date format of the time that this resource was created. */
    created?: Maybe<Scalars['String']>;
    /** The ISO 8601 date format of the time that this resource was edited. */
    edited?: Maybe<Scalars['String']>;
    /**
     * The eye color of this person. Will be "unknown" if not known or "n/a" if the
     * person does not have an eye.
     */
    eyeColor?: Maybe<Scalars['String']>;
    filmConnection?: Maybe<PersonFilmsConnection>;
    /**
     * The gender of this person. Either "Male", "Female" or "unknown",
     * "n/a" if the person does not have a gender.
     */
    gender?: Maybe<Scalars['String']>;
    /**
     * The hair color of this person. Will be "unknown" if not known or "n/a" if the
     * person does not have hair.
     */
    hairColor?: Maybe<Scalars['String']>;
    /** The height of the person in centimeters. */
    height?: Maybe<Scalars['Int']>;
    /** A planet that this person was born on or inhabits. */
    homeworld?: Maybe<Planet>;
    /** The ID of an object */
    id: Scalars['ID'];
    /** The mass of the person in kilograms. */
    mass?: Maybe<Scalars['Float']>;
    /** The name of this person. */
    name?: Maybe<Scalars['String']>;
    /** The skin color of this person. */
    skinColor?: Maybe<Scalars['String']>;
    /** The species that this person belongs to, or null if unknown. */
    species?: Maybe<Species>;
    starshipConnection?: Maybe<PersonStarshipsConnection>;
    vehicleConnection?: Maybe<PersonVehiclesConnection>;
};
// ...
export type Planet = Node & {
    __typename?: 'Planet';
    /** The climates of this planet. */
    climates?: Maybe<Array<Maybe<Scalars['String']>>>;
    /** The ISO 8601 date format of the time that this resource was created. */
    created?: Maybe<Scalars['String']>;
    /** The diameter of this planet in kilometers. */
    diameter?: Maybe<Scalars['Int']>;
    /** The ISO 8601 date format of the time that this resource was edited. */
    edited?: Maybe<Scalars['String']>;
    filmConnection?: Maybe<PlanetFilmsConnection>;
    /**
     * A number denoting the gravity of this planet, where "1" is normal or 1 standard
     * G. "2" is twice or 2 standard Gs. "0.5" is half or 0.5 standard Gs.
     */
    gravity?: Maybe<Scalars['String']>;
    /** The ID of an object */
    id: Scalars['ID'];
    /** The name of this planet. */
    name?: Maybe<Scalars['String']>;
    /**
     * The number of standard days it takes for this planet to complete a single orbit
     * of its local star.
     */
    orbitalPeriod?: Maybe<Scalars['Int']>;
    /** The average population of sentient beings inhabiting this planet. */
    population?: Maybe<Scalars['Float']>;
    residentConnection?: Maybe<PlanetResidentsConnection>;
    /**
     * The number of standard hours it takes for this planet to complete a single
     * rotation on its axis.
     */
    rotationPeriod?: Maybe<Scalars['Int']>;
    /**
     * The percentage of the planet surface that is naturally occurring water or bodies
     * of water.
     */
    surfaceWater?: Maybe<Scalars['Float']>;
    /** The terrains of this planet. */
    terrains?: Maybe<Array<Maybe<Scalars['String']>>>;
};
// ...
export type Starship = Node & {
    __typename?: 'Starship';
    /** The maximum number of kilograms that this starship can transport. */
    cargoCapacity?: Maybe<Scalars['Float']>;
    /**
     * The maximum length of time that this starship can provide consumables for its
     * entire crew without having to resupply.
     */
    consumables?: Maybe<Scalars['String']>;
    /** The cost of this starship new, in galactic credits. */
    costInCredits?: Maybe<Scalars['Float']>;
    /** The ISO 8601 date format of the time that this resource was created. */
    created?: Maybe<Scalars['String']>;
    /** The number of personnel needed to run or pilot this starship. */
    crew?: Maybe<Scalars['String']>;
    /** The ISO 8601 date format of the time that this resource was edited. */
    edited?: Maybe<Scalars['String']>;
    filmConnection?: Maybe<StarshipFilmsConnection>;
    /** The class of this starships hyperdrive. */
    hyperdriveRating?: Maybe<Scalars['Float']>;
    /** The ID of an object */
    id: Scalars['ID'];
    /** The length of this starship in meters. */
    length?: Maybe<Scalars['Float']>;
    /** The manufacturers of this starship. */
    manufacturers?: Maybe<Array<Maybe<Scalars['String']>>>;
    /**
     * The maximum speed of this starship in atmosphere. null if this starship is
     * incapable of atmosphering flight.
     */
    maxAtmospheringSpeed?: Maybe<Scalars['Int']>;
    /**
     * The Maximum number of Megalights this starship can travel in a standard hour.
     * A "Megalight" is a standard unit of distance and has never been defined before
     * within the Star Wars universe. This figure is only really useful for measuring
     * the difference in speed of starships. We can assume it is similar to AU, the
     * distance between our Sun (Sol) and Earth.
     */
    MGLT?: Maybe<Scalars['Int']>;
    /**
     * The model or official name of this starship. Such as "T-65 X-wing" or "DS-1
     * Orbital Battle Station".
     */
    model?: Maybe<Scalars['String']>;
    /** The name of this starship. The common name, such as "Death Star". */
    name?: Maybe<Scalars['String']>;
    /** The number of non-essential people this starship can transport. */
    passengers?: Maybe<Scalars['String']>;
    pilotConnection?: Maybe<StarshipPilotsConnection>;
    /**
     * The class of this starship, such as "Starfighter" or "Deep Space Mobile
     * Battlestation"
     */
    starshipClass?: Maybe<Scalars['String']>;
};
// ...
export type Vehicle = Node & {
    __typename?: 'Vehicle';
    /** The maximum number of kilograms that this vehicle can transport. */
    cargoCapacity?: Maybe<Scalars['Float']>;
    /**
     * The maximum length of time that this vehicle can provide consumables for its
     * entire crew without having to resupply.
     */
    consumables?: Maybe<Scalars['String']>;
    /** The cost of this vehicle new, in Galactic Credits. */
    costInCredits?: Maybe<Scalars['Float']>;
    /** The ISO 8601 date format of the time that this resource was created. */
    created?: Maybe<Scalars['String']>;
    /** The number of personnel needed to run or pilot this vehicle. */
    crew?: Maybe<Scalars['String']>;
    /** The ISO 8601 date format of the time that this resource was edited. */
    edited?: Maybe<Scalars['String']>;
    filmConnection?: Maybe<VehicleFilmsConnection>;
    /** The ID of an object */
    id: Scalars['ID'];
    /** The length of this vehicle in meters. */
    length?: Maybe<Scalars['Float']>;
    /** The manufacturers of this vehicle. */
    manufacturers?: Maybe<Array<Maybe<Scalars['String']>>>;
    /** The maximum speed of this vehicle in atmosphere. */
    maxAtmospheringSpeed?: Maybe<Scalars['Int']>;
    /**
     * The model or official name of this vehicle. Such as "All-Terrain Attack
     * Transport".
     */
    model?: Maybe<Scalars['String']>;
    /**
     * The name of this vehicle. The common name, such as "Sand Crawler" or "Speeder
     * bike".
     */
    name?: Maybe<Scalars['String']>;
    /** The number of non-essential people this vehicle can transport. */
    passengers?: Maybe<Scalars['String']>;
    pilotConnection?: Maybe<VehiclePilotsConnection>;
    /** The class of this vehicle, such as "Wheeled" or "Repulsorcraft". */
    vehicleClass?: Maybe<Scalars['String']>;
};
// ...
export type FilmCharactersEdge = {
    __typename?: 'FilmCharactersEdge';
    /** A cursor for use in pagination */
    cursor: Scalars['String'];
    /** The item at the end of the edge */
    node?: Maybe<Person>;
};

/** A connection to a list of items. */
export type FilmPlanetsConnection = {
    __typename?: 'FilmPlanetsConnection';
    /** A list of edges. */
    edges?: Maybe<Array<Maybe<FilmPlanetsEdge>>>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
    /**
     * A list of all of the objects returned in the connection. This is a convenience
     * field provided for quickly exploring the API; rather than querying for
     * "{ edges { node } }" when no edge data is needed, this field can be be used
     * instead. Note that when clients like Relay need to fetch the "cursor" field on
     * the edge to enable efficient pagination, this shortcut cannot be used, and the
     * full "{ edges { node } }" version should be used instead.
     */
    planets?: Maybe<Array<Maybe<Planet>>>;
    /**
     * A count of the total number of objects in this connection, ignoring pagination.
     * This allows a client to fetch the first five objects by passing "5" as the
     * argument to "first", then fetch the total count so it could display "5 of 83",
     * for example.
     */
    totalCount?: Maybe<Scalars['Int']>;
};

/** An edge in a connection. */
export type FilmPlanetsEdge = {
    __typename?: 'FilmPlanetsEdge';
    /** A cursor for use in pagination */
    cursor: Scalars['String'];
    /** The item at the end of the edge */
    node?: Maybe<Planet>;
};